Modify files of the folder factory_calib

+++ factory_calib/include/dataConvert.hpp
@@ -18,6 +18,7 @@
 #include <string>
 #include <vector>
 
+/*
 static void stdvec2cvmat(const std::vector<std::vector<bool>> &img) {
   std::vector<float> img_uchar;
   for (size_t i = 0; i < img.size(); i++) {
@@ -31,12 +32,14 @@ static void stdvec2cvmat(const std::vector<std::vector<bool>> &img) {
   cv::imshow("dest", dest);
   cv::waitKey();
 }
+*/
 
+/*
 static void stdvec2cvmat(const std::vector<std::vector<float>> &img,
                          cv::Mat &dest) {
   std::vector<float> img_uchar;
-  for (int i = 0; i < img.size(); i++) {
-    for (int j = 0; j < img[0].size(); j++) {
+  for (size_t i = 0; i < img.size(); i++) {
+    for (size_t j = 0; j < img[0].size(); j++) {
       img_uchar.push_back(static_cast<float>(img[i][j]));
     }
   }
@@ -44,12 +47,14 @@ static void stdvec2cvmat(const std::vector<std::vector<float>> &img,
   cv::Mat mat = cv::Mat(img_uchar);
   dest = mat.reshape(1, img.size()).clone();
 }
+*/
 
+/*
 static void stdvec2cvmat(const std::vector<std::vector<int>> &img,
                          cv::Mat &dest) {
   std::vector<float> img_uchar;
-  for (int i = 0; i < img.size(); i++) {
-    for (int j = 0; j < img[0].size(); j++) {
+  for (size_t i = 0; i < img.size(); i++) {
+    for (size_t j = 0; j < img[0].size(); j++) {
       img_uchar.push_back(static_cast<float>(img[i][j]));
     }
   }
@@ -57,14 +62,15 @@ static void stdvec2cvmat(const std::vector<std::vector<int>> &img,
   cv::Mat mat = cv::Mat(img_uchar);
   dest = mat.reshape(1, img.size()).clone();
 }
+*/
 
 static void cvmat2stdvec(const cv::Mat &src,
                          std::vector<std::vector<float>> &img) {
   img.resize(src.rows, std::vector<float>(src.cols));
   std::cout << "img.size(): " << img.size() << std::endl;
   std::cout << "img[0].size(): " << img[0].size() << std::endl;
-  for (int i = 0; i < img.size(); i++) {
-    for (int j = 0; j < img[0].size(); j++) {
+  for (size_t i = 0; i < img.size(); i++) {
+    for (size_t j = 0; j < img[0].size(); j++) {
       img[i][j] = src.at<float>(i, j);
       // img[i][j] = float(src.at<unsigned char>(i,j));
     }
