+++ factory_calib/src/calibration/adas_calib.cpp
@@ -69,7 +69,7 @@ bool CalibrationBoardTool::ChessBoardCalibration(
     color_box.emplace_back(cv::Scalar(80, 80, 235));
     color_box.emplace_back(cv::Scalar(200, 0, 200));
     cv::Mat display_img = image_.clone();
-    for (int i = 0; i < grid_center_points.size(); ++i) {
+    for (size_t i = 0; i < grid_center_points.size(); ++i) {
       cv::Point2f cv_pt;
       cv_pt.x = grid_center_points[i].x;
       cv_pt.y = grid_center_points[i].y;
@@ -299,7 +299,7 @@ bool CalibrationBoardTool::VerticalBoardCalibration(
     color_box.emplace_back(cv::Scalar(80, 80, 235));
     color_box.emplace_back(cv::Scalar(200, 0, 200));
     cv::Mat display_img = image_.clone();
-    for (int i = 0; i < corners.lines_.size(); ++i) {
+    for (size_t i = 0; i < corners.lines_.size(); ++i) {
       std::vector<Eigen::Vector2d> l_pts = corners.lines_[i]->m_line_pts;
       cv::Scalar color = color_box[i % 6];
       for (auto p : l_pts) {
@@ -331,14 +331,14 @@ bool CalibrationBoardTool::VerticalBoardCalibration(
   Mat3d img2board_hmat;
   float grid_size = calibration_inputs.chessboard_grid_size;
   float camera_height = calibration_inputs.camera_height;
-  int mid_corner_num = v_board.lines[1].size();
+  size_t mid_corner_num = v_board.lines[1].size();
   float top_board_height =
       mid_corner_num * grid_size + calibration_inputs.board_height;
   float board_pt_x, board_pt_y;
   for (int i = 0; i < 3; ++i) {
     auto c_line = corners.lines_[i];
-    int pt_num = c_line->m_line_pts.size();
-    for (int j = 0; j < pt_num; ++j) {
+    size_t pt_num = c_line->m_line_pts.size();
+    for (size_t j = 0; j < pt_num; ++j) {
       board_pt_x = (i - 1) * grid_size;
       board_pt_y = top_board_height - corners.pattern_idxs_[i][j] * grid_size -
                    camera_height;
@@ -456,8 +456,8 @@ bool CalibrationBoardTool::CircleBoardCalibration(
     color_box.emplace_back(cv::Scalar(80, 80, 235));
     color_box.emplace_back(cv::Scalar(200, 0, 200));
     cv::Mat display_img = image_.clone();
-    for (int i = 0; i < corners.points.size(); ++i) {
-      for (int j = 0; j < corners.points[i].size(); ++j) {
+    for (size_t i = 0; i < corners.points.size(); ++i) {
+      for (size_t j = 0; j < corners.points[i].size(); ++j) {
         Point2f pt = corners.points[i][j];
         if (pt.x < 1)
           continue;
